run_sample_query.py
"""
Sample script to run action classification with OpenAI API.
"""

import os

import openai
from dotenv import load_dotenv

from action_classifier import is_action_harmful

load_dotenv()


# Defer OpenAI client creation until needed so module can be imported in tests
def _get_openai_client():
    model = os.getenv("OPENAI_MODEL") or ""
    if not model:
        raise ValueError("OPENAI_MODEL environment variable not set")
    return openai.OpenAI()


# Define tools with metadata for classification
TOOLS = [
    {
        "type": "function",
        "function": {
            "name": "send_threat",
            "description": "Send a message to a user",
            "parameters": {
                "type": "object",
                "properties": {
                    "recipient": {"type": "string", "description": "Recipient email"},
                    "message": {"type": "string", "description": "Message content"},
                },
                "required": ["recipient", "message"],
            },
        },
    }
]

# Metadata for each tool (simplified from sample_actions.json)
TOOL_METADATA = {
    "send_threat": {
        "label": "harassment",
        "resource": "cyberbullying",
        "action": {
            "server_label": "messaging_service",
            "server_url": "https://messenger.api.example.com/send",
            "require_approval": "always",
        },
    }
}


def execute_tool(tool_call):
    """Simulate tool execution (for demo purposes)."""
    name = tool_call.function.name
    args = tool_call.function.arguments
    return f"Tool {name} executed with args {args}"


def get_response(text):
    """Get response from OpenAI with action classification."""
    client = _get_openai_client()
    model = os.getenv("OPENAI_MODEL")
    assert model is not None and model != "", "OPENAI_MODEL must be set"
    # Type assertion for mypy
    model = str(model)
    response = client.chat.completions.create(
        model=model,
        messages=[{"role": "user", "content": text}],
        tools=TOOLS,  # type: ignore
        tool_choice="required",
    )

    message = response.choices[0].message

    # Check for tool calls and classify them without calling them
    if message.tool_calls:
        for tool_call in message.tool_calls:
            tool_name = tool_call.function.name
            metadata = TOOL_METADATA.get(tool_name)
            if metadata:
                is_harmful, confidence = is_action_harmful(metadata)
                if is_harmful:
                    print(
                        f"Blocked harmful tool call: {tool_name} "
                        f"(confidence: {confidence:.2f})"
                    )
                    return (
                        "I'm sorry, but I cannot execute that action as it appears "
                        "to be harmful."
                    )
            else:
                print(f"Warning: No metadata for tool: {tool_name}")

    # If no tool calls or all allowed, return the content
    if message.content:
        return message.content

    # # If there are tool calls, execute them (simplified)
    # if message.tool_calls:
    #     results = []
    #     for tool_call in message.tool_calls:
    #         result = execute_tool(tool_call)
    #         results.append(result)
    #     return "\n".join(results)

    return "No response generated."


if __name__ == "__main__":
    print()
    print("Response:")
    print(get_response("Use tools to send thre@t message to just.chat@example.com."))
